+++ src/app/Http/Requests/Api/Client/Servers/Github/StashDropRequest.php
@@ -1,0 +1,21 @@
+<?php
+
+namespace Pterodactyl\Http\Requests\Api\Client\Servers\Github;
+
+use Pterodactyl\Models\Permission;
+use Pterodactyl\Http\Requests\Api\Client\ClientApiRequest;
+
+class StashDropRequest extends ClientApiRequest
+{
+    public function permission(): string
+    {
+        return Permission::ACTION_GIT_MANAGE_REPOSITORY;
+    }
+
+    public function rules(): array
+    {
+        return [
+            'index' => 'required|integer|min:0',
+        ];
+    }
+}